A better way of resetting undolevels
For details see:
https://github.com/AndrewRadev/linediff.vim/commit/bcce15aac63816cf082e4ce69a44566f59271ae7#commitcomment-20298527

Undolevels was a global option up until 7.4.73, so using the
"clear-undo" snippet might reset the undo state globally for an older
Vim. Better to check the version before we commit to doing that.

Also, instead of manually making a change, easier to just restore the
undo state after adding the lines to the buffer.