Remove unnecessary view restoration
Seems like this doesn't do what I wanted it to do. Since the original
buffer is opened in the diff window, the saving and restoration of the
view doesn't really affect the real original window.

Theoretically, it might be possible to switch to the original window,
and make the changes there. But it would be more complicated, and it's a
minor thing. Maybe some time in the future.