Fix bug with multiple update autocommands
Turns out, multiple updates seemed to install multiple "update the
original" autocommands. Hadn't noticed this before, possibly needed a
larger file to work on, or often made less changes with rare updates.

The issue is easy to fix by moving the autocommand installation to its
own function and only calling it on buffer initialization, not every
time the original buffer is updated.